Share Tests.scala between Scala 2 and Scala 3
